requirements.txt
streamlit
pandas
scikit-learn
joblib
matplotlib
Pillow
